+++ Pocket/Pocket/Bridge/FieldController.h
@@ -32,11 +32,11 @@ NS_ASSUME_NONNULL_BEGIN
 -(instancetype)init;
 -(void)initialize;
 
--(nonnull NSArray<Field*>*)getListField:(uint32_t)groupId search:(nonnull const NSString*)search;
+-(nonnull NSArray<Field*>*)getListField:(int64_t)groupId search:(nonnull const NSString*)search;
 -(Stat)persistField:(nonnull const Field*)field;
 -(Stat)delField:(Field*)field;
--(int32_t)sizeFiled:(uint32_t)groupId;
--(nullable Field*)getFiled:(uint32_t)groupId;
+-(int32_t)sizeFiled:(int64_t)groupId;
+-(nullable Field*)getFiled:(int64_t)groupId;
 
 @end
